Get RxCui via Anatomical Therapeutic Chemical Classification System(ATC) code

This is a function to get RxCui via Anatomical Therapeutic Chemical Classification System(ATC) code

getRxCuiViaATC123(df, AtcColName = ATC, cores = 8)

Arguments

df
data.frame, a drug ingredient ATC code of the specific indication

AtcColName
a column name for ATC code of df

cores
number of parallel operation

Differ_ColName
if colum for NHINo of df is not named ATC

Value

A new data.table sample_atc_rxcui with following columns.

RxCui drug RxCui

ATC drug ATC code

Details

Users can obtain additional drug information via other drug integrated by RxCui, such as the contraindications and indications for drugs available in the MED-RT database of the US Veterans Bureau integrated by RxNorm. User inputs the data frame with drug column which encoded in ATC code, then the output will add a column named RxCui in the original input data frame.

Examples

# sample of getting RxCui via ATC.
sample_atc_rxcui <- getRxCuiViaATC123(df = sample_ATC,AtcColName = ATC,cores = 2)
head(sample_atc_rxcui)

#>       ATC  RxCui
#> 1 G04BX06   8120
#> 2 D06AX02   2348
#> 3 C01DA02   4917
#> 4 G03AB01   <NA>
#> 5 A12CC05 142131