Get ingredient RxCui via specific indication

This function provides user to get drug ingredient RxCui of the specific indication may treat.

getRxCuiViaMayTreatTest1(strmaytreat)

Arguments

strmaytreat
user can input an indication name

Value

A new data.table maytreat_List with following columns.

RxCui drug RxCui

Name drug name

MinConcept.Id indication Id

MinConcept.Name indication name

Details

The data source is from the U.S. Veterans Administration's MED-RT database. Since RxNorm integrates the US Veterans Administration MED-RT drug vocabulary and provides API (/rxclass/relaSources) that mediates RxCui into the drug vocabulary. Therefore, we establishes a function through this API, so that users can obtain complete contraindications, indications and other information in the MED-RT vocabulary through the drug RxCui. In addition to drug contraindications and indications for RxNorm coding, this study also uses web crawler technology to crawl information on contraindications and indications for all drugs in the MED-RT data sheet. Establish a function for finding drugs from contraindications and indications.

Examples

# sample of searching an indication esophagitis may treat.
Esophagitis_List <- getRxCuiViaMayTreatTest1("esophagitis")
# sample of getting drug ingredient RxCui list of esophagitis.
head(Esophagitis_List)

#>      RxCui                 Name MinConcept.Id MinConcept.Name
#> 1: 1792108    Omeprazole Sodium       D004941     Esophagitis
#> 2:  236486 Omeprazole magnesium       D004941     Esophagitis
#> 3:    7646           Omeprazole       D004941     Esophagitis
#> 4:   10156           Sucralfate       D004941     Esophagitis
#> 5: 1487143    sucralfate malate       D004941     Esophagitis
#> 6:   17128         lansoprazole       D004941     Esophagitis