Get Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui via RxCui

This function provides user to get Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui via ingredient RxCui.

getSBDRxCuiViaRxCui123(df, RxCuiColName = RxCui, cores = 8)

Arguments

df
data.frame, a drug ingredient RxCui of the specific indication

RxCuiColName
a column name for RxCui of df

cores
number of parallel operation

Value

A new data.table maytreat_SBD_SCD_List with following columns.

RxCui drug RxCui

Name drug name

MinConcept.Id indication Id

MinConcept.Name indication name

SBD.rxcui drug SBD RxCui

Details

After obtaining the data.frame from getRxCuiViaMayTreat() and getRxCuiViaMayPrevent(), user can further obtain a commercially available pharmaceutical brand and clinical drug contained these drug ingredients. The content is to add a column named SBD.rxcui in the original input data frame, stores the result of converting ingredient to brand drug and clinical drug containing a specific drug component.

Examples

# sample of getting SBD or SCD Rxcui list of esophagitis.
Esophagitis_SBD_SCD_List <- getSBDRxCuiViaRxCui123(df = Esophagitis_List[3,],RxCuiColName = RxCui,cores = 2)
head(Esophagitis_SBD_SCD_List)

#>    RxCui       Name MinConcept.Id MinConcept.Name SBD.rxcui
#> 1:  7646 Omeprazole       D004941     Esophagitis   2262026
#> 2:  7646 Omeprazole       D004941     Esophagitis   2003656
#> 3:  7646 Omeprazole       D004941     Esophagitis    198051
#> 4:  7646 Omeprazole       D004941     Esophagitis    199119
#> 5:  7646 Omeprazole       D004941     Esophagitis    389171
#> 6:  7646 Omeprazole       D004941     Esophagitis    389172